refactor: router config
