+++ config/router/src/custom-bases.ts
@@ -1,0 +1,5 @@
+import { Token } from '@sushiswap/currency'
+
+export const CUSTOM_BASES: {
+  [chainId: number]: { [tokenAddress: string]: Token[] }
+} = {}
